start consolidate threejs segment funcitons
